requirements.txt
opencv-python
pyserial
flask